+++ frontend/src/pages/Management/PlayerManagement.vue
@@ -2,17 +2,7 @@
     <div class="container-fluid">
         <div class="row mb-3 mt-3">
             <div class="col-lg-12">
-                <h1>Player Groups Management</h1>
-                <p class="mb-0">
-                    Login URL:
-                    <a :href="`${backendHost}/login/${loginNames.managed}`">
-                        {{ backendHost }}/login/{{ loginNames.managed }}
-                    </a>
-                </p>
-                <p class="small text-muted">
-                    This login URL does not require ESI scopes. It disables groups for the player account
-                    if the "Deactivate Groups" feature is enabled, unless their account status is "managed".
-                </p>
+                <h1>Player Management</h1>
             </div>
         </div>
 
@@ -28,7 +18,7 @@
                         <a v-for="char in searchResult"
                            class="list-group-item list-group-item-action"
                            :class="{ active: playerId === char.player_id }"
-                           :href="`#PlayerGroupManagement/${char.player_id}`">
+                           :href="`#PlayerManagement/${char.player_id}`">
                             {{ char.character_name }}
                         </a>
                     </div>
@@ -36,13 +26,13 @@
                 <div class="card border-secondary mb-3" >
                     <h4 class="card-header">
                         Players
-                        <span class="hdl-small">status = managed</span>
+                        <span class="hdl-small">status = manually managed</span>
                     </h4>
                     <div class="list-group">
                         <span v-for="managedPlayer in players">
                             <a class="list-group-item list-group-item-action"
                                :class="{ active: playerId === managedPlayer.id }"
-                               :href="`#PlayerGroupManagement/${managedPlayer.id}`">
+                               :href="`#PlayerManagement/${managedPlayer.id}`">
                                 {{ managedPlayer.name }} #{{ managedPlayer.id }}
                             </a>
                         </span>
@@ -50,36 +40,43 @@
                 </div>
             </div>
             <div class="col-lg-8">
-                <div class="card border-secondary" >
-                    <h3 class="card-header">Groups</h3>
-                    <div v-if="playerData" class="card-body">
-                        <h4>
-                            <span v-if="h.hasRole('user-admin')">
-                                <a :href="`#UserAdmin/${playerData.id}`" title="User Administration">
-                                    {{ playerData.name }} #{{ playerData.id }}</a>
-                            </span>
-                            <span v-if="!h.hasRole('user-admin')">{{ playerData.name }} #{{ playerData.id }}</span>
-                        </h4>
+                <div v-cloak v-if="playerData" class="card border-secondary border-bottom-0">
+                    <h3 class="card-header">
+                        {{ playerData.name }} #{{ playerData.id }}
+                    </h3>
+
+                    <div class="card-body">
                         <p>
-                            Status: {{ playerData.status }}
-                            <a class="btn btn-info nc-btn-xs ms-1" href=""
-                               v-on:click.prevent="h.showCharacters(playerData.id)">
+                            Status:
+                            <span v-if="playerData.status === 'managed'">manually managed</span>
+                            <span v-else>standard</span>,
+
+                            <span v-if="h.hasRole('user-admin')">
+                                <a :href="`#UserAdmin/${playerData.id}`">User Administration</a>,
+                            </span>
+
+                            <a class="" href="" v-on:click.prevent="h.showCharacters(playerData.id)">
                                 Show characters
                             </a>
-                        </p>
-                        <p v-if="playerData.status === 'standard'" class="text-warning">
-                            The status of this player is not "managed", manual changes will
-                            be overwritten by the automatic group assignment.
                         </p>
 
                         <hr>
 
-                        <h5>Account Status</h5>
-                        <p class="text-warning">
-                            All groups will be removed from the player account when the status is changed!
+                        <h4>Account Status</h4>
+                        <p>
+                            <strong>Standard</strong>: Automatic group assignments, requires valid ESI token for
+                            <a href="#SystemSettings/Features">Groups Deactivation</a> feature.<br>
+                            <strong>Manually managed</strong>: No automatic group assignments, does not require a valid ESI
+                            token.<br>
+                            See also Documentation
+                            <a :href="settings.repository + '/blob/main/doc/Documentation.md#account-status'"
+                               class="external" target="_blank" rel="noopener noreferrer">Account Status</a>
+                            and
+                            <a :href="settings.repository + '/blob/main/doc/Documentation.md#group-deactivation'"
+                               class="external" target="_blank" rel="noopener noreferrer">Group Deactivation</a>.
                         </p>
                         <div class="input-group mb-1">
-                            <label class="input-group-text" for="userAdminSetStatus">status</label>
+                            <label class="input-group-text" for="userAdminSetStatus">Status</label>
                             <select class="form-select" id="userAdminSetStatus"
                                     v-model="playerData.status"
                                     v-on:change="setAccountStatus()">
@@ -87,14 +84,24 @@
                                 <option value="managed">manually managed</option>
                             </select>
                         </div>
+                        <p class="text-warning">
+                            All groups will be removed from the player account when the status is changed!
+                        </p>
+
+                        <hr>
+
+                        <h4 class="mb-0">Groups</h4>
+                        <p class="mt-2 mb-0" v-if="playerData.status === 'standard'">
+                            Note: The status of this player is not "manually managed", manual changes to
+                            group membership will be overwritten by the automatic group assignment.
+                        </p>
                     </div>
                 </div>
-
                 <admin v-cloak v-if="playerId" ref="admin"
                        :contentType="'groups'" :typeId="playerId" :type="'Player'"
+                       :cardClass="'border-top-0'" :cardBodyClass="'pt-0'"
                        v-on:activePlayer="playerData = $event"></admin>
-
-            </div>
+            </div> <!-- col -->
         </div>
     </div>
 </template>
@@ -122,12 +129,12 @@ export default {
     data() {
         return {
             h: new Helper(this),
+            settings: toRef(this.store.state, 'settings'),
             player: toRef(this.store.state, 'player'),
             loginNames: Data.loginNames,
             players: [],
             playerId: null, // current player
             playerData: null, // current player
-            backendHost: null,
             searchResult: [],
         }
     },
@@ -137,9 +144,6 @@ export default {
 
         this.getPLayers();
         this.setPlayerId();
-
-        // login URL for managed accounts
-        this.backendHost = Data.envVars.backendHost;
     },
 
     watch: {
@@ -160,6 +164,9 @@ export default {
 
         setPlayerId() {
             this.playerId = this.route[1] ? parseInt(this.route[1], 10) : null;
+            if (this.playerId === null) {
+                this.playerData = null;
+            }
         },
 
         onSearchResult(result) {